add test for delete profile and fix bugs

+++ apps/files-microservice/src/features/pictures/application/use-cases/delete.profile.use.case.ts
@@ -16,7 +16,11 @@ export class DeleteProfileUseCase
 
   async execute(command: DeleteProfileCommand): Promise<void> {
     const avatar = await this.avatarsRepository.getAvatarInfo(command.userId);
-    await this.avatarsRepository.deleteAvatarInfo(command.userId);
-    await this.fileStorage.deleteImage(avatar.bucket, avatar.key);
+    if (avatar) {
+      await this.avatarsRepository.deleteAvatarInfo(command.userId);
+      await this.fileStorage.deleteImage(avatar.bucket, avatar.key);
+    } else {
+      return;
+    }
   }
 }
